:truck: Move convertArrayMethod into array namespace
