+++ src/util/convert.js
@@ -1,5 +1,4 @@
 import lodashFpConvert from './lodashFpConvert'
-import toArray from 'lodash/toArray'
 import update from '../object/update'
 
 /**
@@ -33,20 +32,3 @@ export default convert
  * @private
  */
 export const convertLodashFp = fn => convert(lodashFpConvert(fn))
-
-/**
- * Converts an Array method.
- * @function
- * @memberof util
- * @param {string} method Array method name.
- * @return {function} Returns the wrapped function.
- * @see {@link util.convert|convert} for more information.
- * @see {@link https://lodash.com/docs#toArray|lodash.toArray} for more information.
- * @since 0.1.14
- * @private
- */
-export const convertArrayMethod = method => convert((array, ...args) => {
-  const newArray = toArray(array)
-  newArray[method](...args)
-  return newArray
-})
